refactor: split region map visibility and marker parts (#3878)
Decompose shared visibility helpers into fog/transport vs extraction
parts; split civilian/fleet unit markers and capital settlement paint
into dedicated library parts for Phase 3 flame map modularization.
