+++ SPEC/program/repo-and-packages.md
@@ -22,7 +22,7 @@ Monorepo layout:
 - **`features/game/flame/caches/`** — shared Flame asset decode helper (`AssetImageCache`) and map marker icon caches (town, port, fleet, civilian, resource, province label). Public barrel: `caches/caches.dart`.
 - **`features/game/flame/map_area/`** — `GameMapAreaBackground` and other render-only map chrome. Public barrel: `map_area/map_area.dart`.
 - **`features/game/flame/map_state/`** — `GameMapArea` widget stack, selection, province actions, turn-feed hooks, and draft projections. Public barrel: `map_state/map_state.dart`.
-- **`features/game/flame/region_map/`** — `RegionMapComponent` Flame render stack and viewport helpers. `region_map_component.dart` library parts split shared constants/plate layout (`region_map_component_shared.dart`) from visibility, fog, label, and overlay helpers (`region_map_component_shared_visibility*.dart`); unit marker paint (`region_map_component_render_markers_units.dart`); capital/town/port paint (`region_map_component_render_markers_settlements_towns.dart`) from warp-zone edges (`region_map_component_render_markers_settlements_warp.dart`); province label compute vs paint (`region_map_component_render_political_labels_province_compute.dart`, `region_map_component_render_political_labels_province_paint.dart`). Public barrel: `region_map/region_map.dart`.
+- **`features/game/flame/region_map/`** — `RegionMapComponent` Flame render stack and viewport helpers. `region_map_component.dart` library parts split shared constants/plate layout (`region_map_component_shared.dart`) from visibility modes (`region_map_component_shared_visibility.dart`), fog/transport helpers (`region_map_component_shared_visibility_fog_transport.dart`), extraction disc helpers (`region_map_component_shared_visibility_extraction.dart`), and label/halos (`region_map_component_shared_visibility_*.dart`); civilian vs fleet unit marker paint (`region_map_component_render_markers_units_civilian.dart`, `region_map_component_render_markers_units_fleet.dart`); capital rings (`region_map_component_render_markers_settlements_capitals.dart`), town/port paint (`region_map_component_render_markers_settlements_towns.dart`), and warp-zone edges (`region_map_component_render_markers_settlements_warp.dart`); province label compute vs paint (`region_map_component_render_political_labels_province_compute.dart`, `region_map_component_render_political_labels_province_paint.dart`). Public barrel: `region_map/region_map.dart`.
 - **`features/game/flame/tilesets/`** — terrain Wang tilesets, transport overlay tilesets, and connectivity masks. Public barrel: `tilesets/tilesets.dart`.
 - **`features/game/flame/controls/`** — in-game map shell chrome: top/tab bar ([GameMapControls](../../app/lib/features/game/flame/controls/game_map_controls.dart)), corner tool row ([GameMapCornerControls](../../app/lib/features/game/flame/controls/game_map_corner_controls.dart)), and empire left rail ([GameMapEmpireLeftRail](../../app/lib/features/game/flame/controls/game_map_empire_left_rail.dart)). Public barrel: `controls/controls.dart`.
 - **`features/game/flame/overlays/`** — map and session overlays/dialogs: province detail hosts ([GameMapProvinceDetailSidePanel](../../app/lib/features/game/flame/overlays/game_map_province_detail_side_panel.dart), [GameMapNarrowDetailOverlaySlot](../../app/lib/features/game/flame/overlays/game_map_narrow_detail_overlay.dart)), turn confirmation/processing dialogs, victory overlay, exit confirm, and debug console panel. Public barrel: `overlays/overlays.dart`.
