+++ app/lib/features/game/flame/region_map/region_map_component_render_markers_settlements_towns.dart
@@ -2,29 +2,6 @@
 part of 'region_map_component.dart';
 
 extension _CtRegionMapRenderMarkersSettlementsTowns on CtRegionMapComponent {
-  void _paintCapitals(Canvas canvas) {
-    final fill = Paint()..style = PaintingStyle.fill;
-    final stroke = Paint()
-      ..style = PaintingStyle.stroke
-      ..strokeWidth = kCapitalMarkerRingStrokeWidth
-      ..color = Colors.black;
-    for (final cap in region.capitalMarkers) {
-      final cell = region.cellAt(cap.x, cap.y);
-      if (regionMapSkipPointMarkerOnCell(
-        playerConstrainedVisibility:
-            visibilityMode == CtMapVisibilityMode.playerConstrained,
-        cellVisibility: _visibilityForTerrain(cell),
-      )) {
-        continue;
-      }
-      final cx = cap.x * cellSize + cellSize / 2;
-      final cy = cap.y * cellSize + cellSize / 2;
-      fill.color = _kMapSelectionGold;
-      canvas.drawCircle(Offset(cx, cy), 6, fill);
-      canvas.drawCircle(Offset(cx, cy), 6, stroke);
-    }
-  }
-
   void _paintTowns(Canvas canvas) {
     if (!townIconCache.isLoaded) return;
